ai_interviewer - local.ai/backend/requirements.txt
Flask
psycopg2-binary
requests
openai
python-dotenv
